requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
python-dotenv==1.0.0
azure-identity==1.17.1
msgraph-sdk==1.4.0
pydantic-settings==2.3.4